pyproject.toml
[tool.poetry]
name = "p2p-platfor-project"
version = "0.1.0"
description = "Back p2p traffic platform"
authors = ["CodeArtistanse"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
tortoise-orm = "^0.19.0"
aiogram = "^3.15.0"
aiohttp = "^3.8.1"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.8"

[build-system]
requires = ["poetry>=1.3.0"]
build-backend = "poetry.core.masonry.api"
